IPC FW3.0 | build project | implementation of event library; Description: modify code of solve warning of type; Risk: low; Affected Module: sys_server; Test: compiled on amba; Code Review: no;
git-svn-id: https://192.168.0.21/svn/ipc_fw3.x_core/branches/20131125@1341 dd879d07-adaa-4757-b7d0-008d00b626b4

+++ modules/gmi/event_center/source/event_process_light.cpp
@@ -145,7 +145,7 @@ void_t* EventProcessLight::TimerEntry()
 		
 		m_OperationLock.Lock( TIMEOUT_INFINITE );
 		if((GetTriggedTime() > 0)
-			&&((CurrTime > (GetTriggedTime()+GetDelayTime())) || (GetTriggedTime() > CurrTime)))
+			&&((CurrTime > (GetTriggedTime()+GetDelayTime())) || (GetTriggedTime() > (uint32_t)CurrTime)))
 		{	
 			printf("GMI_ALARM_MODE_LIGHT close[%u]\n", (uint32_t)CurrTime);
             Result = GMI_BrdSetAlarmOutput( GMI_ALARM_MODE_LIGHT, 0, 0 );
